+++ inc/c_abi/Tecs_entity_view.h
@@ -11,11 +11,11 @@ extern "C" {
 
 typedef struct tecs_entity_view_t {
     const void *storage;
-    size_t start_index;
-    size_t end_index;
+    uint64_t start_index;
+    uint64_t end_index;
 } tecs_entity_view_t;
 
-TECS_EXPORT size_t Tecs_entity_view_storage_size(const tecs_entity_view_t *view);
+TECS_EXPORT uint64_t Tecs_entity_view_storage_size(const tecs_entity_view_t *view);
 TECS_EXPORT const tecs_entity_t *Tecs_entity_view_begin(const tecs_entity_view_t *view);
 TECS_EXPORT const tecs_entity_t *Tecs_entity_view_end(const tecs_entity_view_t *view);
